refactor : 변경된 게시판 명으로 수정
