--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/board/BuysellBoardMapper.java
@@ -1,7 +1,0 @@
-package com.codestates.main07.buysellBoard.board;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface BuysellBoardMapper {
-}
