--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/board/BuysellBoardRepository.java
@@ -1,8 +1,0 @@
-package com.codestates.main07.buysellBoard.board;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BuysellBoardRepository extends JpaRepository<BuysellBoard, Long> {
-}
