--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/board/BuysellBoardServiceImpl.java
@@ -1,8 +1,0 @@
-package com.codestates.main07.buysellBoard.board;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class BuysellBoardServiceImpl {
-}
