--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/board/dto/BuysellBoardUpdateDto.java
@@ -1,6 +1,0 @@
-package com.codestates.main07.buysellBoard.board.dto;
-
-
-public class BuysellBoardUpdateDto {
-
-}
