--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/comment/BuysellBoardCommentMapper.java
@@ -1,7 +1,0 @@
-package com.codestates.main07.buysellBoard.comment;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface BuysellBoardCommentMapper {
-}
