--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/comment/BuysellBoardCommentRepository.java
@@ -1,8 +1,0 @@
-package com.codestates.main07.buysellBoard.comment;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BuysellBoardCommentRepository extends JpaRepository<BuysellBoardComment, Long> {
-}
