--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/comment/BuysellBoardCommentServiceImpl.java
@@ -1,7 +1,0 @@
-package com.codestates.main07.buysellBoard.comment;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BuysellBoardCommentServiceImpl {
-}
