--- Server/main07/src/main/java/com/codestates/main07/buysellBoard/comment/dto/BuysellBoardCommentUpdateDto.java
@@ -1,6 +1,0 @@
-package com.codestates.main07.buysellBoard.comment.dto;
-
-
-public class BuysellBoardCommentUpdateDto {
-
-}
