+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/board/MarketBoardController.java
@@ -1,4 +1,4 @@
-package com.codestates.main07.buysellBoard.board;
+package com.codestates.main07.marketBoard.board;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RestController;
 @Validated
 @RequestMapping("/buysellBoards")
 @RestController
-public class BuysellBoardController {
+public class MarketBoardController {
 }
