+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/board/MarketBoardMapper.java
@@ -1,0 +1,7 @@
+package com.codestates.main07.marketBoard.board;
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface MarketBoardMapper {
+}
