+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/board/MarketBoardRepository.java
@@ -1,0 +1,8 @@
+package com.codestates.main07.marketBoard.board;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MarketBoardRepository extends JpaRepository<MarketBoard, Long> {
+}
