+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/board/MarketBoardService.java
@@ -1,0 +1,12 @@
+package com.codestates.main07.marketBoard.board;
+
+import com.codestates.main07.marketBoard.board.dto.BuysellBoardResponseDto;
+
+import java.util.List;
+
+public interface MarketBoardService {
+
+    public List<MarketBoard> buysellBoardList();
+
+    public BuysellBoardResponseDto viewBuysell (Long id);
+}
