+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/board/dto/MarketBoardUpdateDto.java
@@ -1,0 +1,6 @@
+package com.codestates.main07.marketBoard.board.dto;
+
+
+public class BuysellBoardUpdateDto {
+
+}
