+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/comment/MarketBoardCommentMapper.java
@@ -1,0 +1,7 @@
+package com.codestates.main07.marketBoard.comment;
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface MarketBoardCommentMapper {
+}
