+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/comment/MarketBoardCommentService.java
@@ -1,0 +1,12 @@
+package com.codestates.main07.marketBoard.comment;
+
+import com.codestates.main07.marketBoard.board.dto.BuysellBoardResponseDto;
+
+import java.util.List;
+
+public interface MarketBoardCommentService {
+
+    public List<MarketBoardComment> buysellBoardCommentList();
+
+    public BuysellBoardResponseDto viewBuysellComment (Long id);
+}
