+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/comment/MarketBoardCommentServiceImpl.java
@@ -1,0 +1,7 @@
+package com.codestates.main07.marketBoard.comment;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class MarketBoardCommentServiceImpl {
+}
