+++ Server/main07/src/main/java/com/codestates/main07/marketBoard/comment/dto/MarketBoardCommentSaveDto.java
@@ -1,0 +1,6 @@
+package com.codestates.main07.marketBoard.comment.dto;
+
+
+public class BuysellBoardCommentSaveDto {
+
+}
